Don't attempt to merge root and shard listings in proxy
See https://etherpad.openstack.org/p/deep-container-listing

The existing merge code was insufficient for merging records of the
same object that might have differing relative meta and content-type
timestamps.

The merging was only necessary for shard ranges in CREATED state that
had not yet been cleaved but might have recent objects updates that
are not in the root. The proxy no longer gets those shard ranges from
the container serve so merging is not required.

Change-Id: Ied70aea3d0b077af484d7e1bb95d929a7fa927b0