Define singletons to represent min and max shard range bounds
Previously the min lower bound and max upper bound of a ShardRange
were both represented by the empty string. This led to complex
conditional statements. For example:

- when both lower and upper were unbounded then it was not possible to
  simply check that lower != upper

- it was not possible to simply check if one shard range's upper was
  less than another, because the empty string is always less than any
  other string.

Those kinds of tests therefore needed confusing extra logic to first
test if the upper existed at all before making the expected
comparison.

With this patch the min lower and max upper bounds are represented by
singleton objects that have the property that ShardRange.MIN is less
than any other name and ShardRange.MAX is greater than any other
name. Consequently it is now safe to write:

  if sr1.upper > sr2.upper:
      print 'sr1 upper is definitely greater than sr2.upper'

even when sr1.upper is unbounded.

ShardRange.MIN and ShardRange.MAX both use the empty string as their
string representation so they can still be serialized in the same way
as before. Also, ShardRange.lower and ShardRange.upper can still
be set with the empty string but internally this is converted to the
appropriate singleton.

Also, add properties and setters for ShardRange lower and upper
attributes to implement checking.

Change-Id: I00a45b5a29531cb06df28b18e88e38ab6285b5b9