Simplify proxy-server getting sharded listings
Ensure that the container-server GET response for shard ranges always
returns a list that covers the entire namespace. This allows
simplification of the proxy-server side GET code.

While in sharding state the container may not yet have an entire
namespace worth of shard ranges so when necessary a filler shard range
is added on the container-server side to cover any as yet uncleaved
namespace.

Drive-by fix for container listing limit param not being cast to int
before being used in calculations.

Adds new probe test specifically to test listings with params during
sharding.

Change-Id: Ib38d75e04008509f4934bddf885fe030e56dfad0