Solve the zombie process problem of Auditor
As the bug 1743310 reported,if we list the swift processes ,we will 
see a zombie process every one minute.The zombie processes numbers may 
be more than one. 

The related code as follows: swift/obj/auditor.py:386~397

        while pids:
            pid = os.wait()[0]
            # ZBF scanner must be restarted as soon as it finishes
            # unless we're in run-once mode
            if self.conf_zero_byte_fps and pid == zbf_pid and \
               len(pids) > 1 and not once:
                kwargs['device_dirs'] = override_devices
                # sleep between ZBF scanner forks
                self._sleep()
                zbf_pid = self.fork_child(zero_byte_fps=True, **kwargs)
                pids.add(zbf_pid)
            pids.discard(pid)

The pids list includes a zbf pid and one or more pids with mode 
(ALL - parallel, objectXXX).If the zbf process is the last one 
finished，then the second zbf process will not be forked.Conversely， 
the zbf process will be forked again,before the fork procedure,the main 
process will sleep for self.interval seconds,30 seconds by default,
during the self.interval time,if a non zbf process finished,this process 
will become a zombie,because the main process is in sleep.

In this solution,we move the sleep from whileloop to the subprocess,and
the main process will not be blocked in whileloop,so the subprocess
will be recovered in time.

Closes-Bug: #1743310

Change-Id: I61c766aa2a1c4bad0247a44a8e78ef38d9f3ae47