Add a note about the cost of COPY for setting metadata
The pointer to using COPY to the same object as a mechanism to set only
a subset of the metadata, it does not mention that doing so results in
a full copy of the object in question on the backend.

Add a note so it's clear that there is a tradeoff involved.

Change-Id: I0c20a4909a6c3ff672f753d26cb9fb2f5f33d1f4