object-server can 409 in response to x-if-delete-at
Previously, if the expirer had a stale work item (because the object
was overwritten or deleted, or some other process handled the delete),
then it would keep retrying for reclaim_age, but every time it'd get
back a 412.

Now, have the object-server be smart enough to say, "I have more recent
information than you" and let the expirer accept that as success.

Change-Id: I0a94482ed16cb30ce79074e053e6177fe97bcaa9