Use backend header rather than param to GET shard record type
X-Backend-Record-Type=shard is used when shard ranges are being PUT
to a container. For consistency, use the same with GET requests for
shard ranges, instead of a query param items=shard.

This removes need for gatekeeper to filter out items=* params
from client supplied query strings. That in turn avoids the
slim possibility of breaking third party middleware that
used items= params.

The item param support will be removed in a separate patch.

Change-Id: Idda136b7a896f572c56943f0ae29321fa73af935