simplify shrinking
When a donor/acceptor pair are found for shrinking, don't update their
timestamps and namespaces in the root db immediately but wait for the
shrinking shard to send updates back to root. That way the root shard
range table only changes once the donor has replicated to the
acceptor.

Before, the acceptor shard range namespace was expanded in the root
shard range db *before* the donor objects had cleaved to it, so there
was a window of time when a listing would have 'lost' the donor
objects.

Also, send a single PUT request to donor to both update its shard
ranges and set its Scan-Done sysmeta. This requires the container
server accepting sysmeta updates with shard range updates.

Change-Id: I3fddaf1f3dd24abfb4f3011d792580719ffc27cd