Don't lose objects when db is rsync replicated during sharding
Before, if an rsync replication updated a hash.db that was partially
cleaved and objects were added in the already cleaved namespace, the
hash.db would be deleted when cleaving completed and the new object(s)
would be lost.

With this patch the sharder will check that the hash.db object table
has the same max_row at start and end of cleaving, and if not will
repeat cleaving. That way it detects if an rsync replication adds
rows.

Note that usync replication during sharding will add rows to the
_shard.db and these are handled as misplaced objects.

Change-Id: I0aa2e2b077cd3ae1fc73985051c0302230f5c603