Refactor expirer unit tests
In expirer's unit tests, fake InternalClient classes are defined
and its instance simulates expirer's task queue behaviors.

To make review for expirer's task queue update patch [1] easy,
this patch refactors the implementation of the fake InternalClient
classes. In this patch, unit tests are refactored by the following
two approaches:
    #1: Summarizing duplicated fake InternalClient implementation
    #2: Make task account name variable
The #2 approach is for multiple task accounts in [1].

The patch [1] will be rebased after this patch merged.

[1]: https://review.openstack.org/#/c/517389

Change-Id: I10a7151cfdd43460ad38c47f672d3c31b77e7990