Remove special case for mismatched last cleaved pointer
Remove cleaving code for special handling when
X-Container-Sysmeta-Last-N (a.k.a 'last_upper') did not align with a
shard range upper (i.e. shard ranges must have changed since last
batch of cleaving). It wasn't clear how and why the special case code
worked.

If last_upper is aligned with a shard range bound then cleaving will
start with the shard range whose lower == last_upper.

If last_upper is for some reason not aligned with a shard range bound
then cleaving will start with the shard range that *contains*
last_upper. This may mean that some objects are cleaved twice from
that shard range, but that's ok, this is the uncommon case.

Change-Id: I72884f4c8062532dd8e3e1defa32f8bbeeeb47f0