Make find_shard_ranges cope with misplaced objects
Before if a shard db had misplaced objects (i.e. names outside of the
shard namespace) then find_shard_ranges() could blow up.

Now, the final shard range found is always limited to the shard's
upper namespace bound even if there happen to be misplaced objects
beyond that bound. The final shard range object count may be
distorted, but that is temporary until cleaving completes, and the
distortion is same as that of the original container object count.

Change-Id: Iad9018791cd09d86c31308f0911f809572e088e1