Remove some cruft from ratelimit tests
The tests were carefully setting up a mock for the 'http_connect'
function in the ratelimit module, but there is no such function
imported by the ratelimit module.

As far as I can tell, this has been the case since the ratelimit
middleware first appeared in 72d40bd (Mon Oct 4 14:11:48 2010 -0700).

Change-Id: If047184c6435aa1647050f50b499dc9feff4318d