ratelimit: ignore requests with invalid API versions
If you've got things like domain_remap, swift3, or other such
middlewares in your pipeline, you wind up with requests that aren't of
the form "/v1/<account>/<container>/<object>". When encountering such
an oddball request, it's not useful to call get_account_info() on the
second path component since it's probably not an account.

This commit makes the ratelimit middleware skip requests that don't
start with either "/v1" or "/v1.0". The requests will still be
handled, but they won't be rate-limited.

Change-Id: I9980cd0e902610ac99d13a502ae955bca2d99df3
Closes-Bug: 1669888
Closes-Bug: 1695273