Reduce scope: do not merge misplaced objects into container listing
Before, while a container was sharding, any misplaced objects
(i.e. updates received since sharding started, for ranges not yet
found) would be merged with the listing from the retired db in an
attempt to give the most up to date listing. However, the merging code
was flawed as it did not respect the data, content-type and metadata
timestamps when merging, but erroneously assumed that all misplaced
objects were newer than those in the retired db.

This patch removes the flawed merge code and consequently listings of
as yet uncleaved ranges will not include recently received updates
that are in the misplaced objects table.

Misplaced objects will be included when cleaving a shard range and will
then appear in the concatenated listing that is constructed by the
proxy server.

Also removes an 'expectedFailure' unit test that covered merging
lists from two dbs.

Change-Id: If58a2e943ccca2e6898dce7c9a2883e870d9cdbb