Revert to root container if object update redirection loops
The async pending file now maintains a list of previously used
redirect locations. If a redirect is received to a location that has
been previously used, then it is ignored and the next attempt of the
async update is sent to the root container.

Note that the cycle could repeat, but at least there is now a
predictable fall back to the root container which is assumed to
eventually have the correct redirect location.

Add more unit tests for updater receiving a bad Location header.

Change-Id: Ie4263b9f700c7a97b72e9cdf95c4854690b9c878