Update root with shard's own shard range even once sharded
Once a shard has sharded, it must update the root to indicate that
it's own shard range is deleted and the sub-shards are now
active. Before this happened when cleaving completed, but if those
requests failed then there was no further attempt.

Now, repeatedly update root even when shard has been sharded, to be
sure the root gets the state changes. This restores behaviour removed
in [1], but retains the update at completion of cleaving which is
timely in case the shard then remains in sharding state.

Note that each repeated update uses same timestamp so updates will be
idempotent.

[1] Related-Change: I0aa2e2b077cd3ae1fc73985051c0302230f5c603

Change-Id: Ied777f7dfa35e790d9cae664f7fc1330854a7660