Always send redirected object updates to all container nodes
Before, if an update request received success from some nodes and
redirect from others, the subsequent redirected update would only be
sent to a subset of shard container nodes whose indexes were not
previously recorded as successful.

Now, any redirect will cause memory of previous successes to be erased
so that all shard container nodes get sent the redirected update.

Also, tweak the stats logging so that redirects are not recorded as
failures, but are always recorded.

Also, refactor and try to minimise the changes in updater.py
w.r.t. master branch.

Change-Id: I556a34d501df9a9832b1ece05a18f8696b72650e