Make object server persist redirect location in async pending
Before, if the object server's container update gets a redirect
response then an async pending file was written, but the redirect
location was not included. So the object updater's first request would
be to the original location and presumably it would then get the same
redirect response.

Now, the async pending file includes any redirect location, so the
object updater can use that in its first request.

Change-Id: I0a0c391fdf5aeafc0644074b61ae8cce882ebc88