Fix merge_shard_ranges to preserve existing bounds when created_at is equal
When merging a new shard range with existing, if the created_at timestamps
are equal then the lower and upper bounds of the existing should always be
kept.

Before, the bounds of the new item might be merged if the new item also had
metadata or state changes at newer meta or state timestamps.

Change-Id: I221ac287b794b955c9f5def7034187c00a37586e