Move shrinking shard ranges to a 'shrinking' state
Once a shard range has been identified as a donor for shrinking it
should not be used as an acceptor during a subsequent cycle, so change
its state to reflect that it is shrinking.

A shrinking range should still be treated the same as an active range
for the purposes of container listings, until it is deleted
simultaneously with the acceptor taking over its namespace in the root
shard ranges table.

To avoid bleeding details of container backend states to the proxy,
introduce a state alias 'listing' which the proxy can use to
parameterize a GET request for all shard ranges that should be used
when constructing a listing.  In the backend this is mapped to
'active,shrinking'.

Change-Id: If1e73b330fe1adecd8539eee976a1aa3d0554261