Retry creating shard containers until successful
When shard ranges are first found, a corresponding container is
created by a PUT.

Before, if this PUT failed, it was not repeated. That was ok because
eventually the shard range would be cleaved and that would create the
container. Or, an object update redirected to that shard might
autocreate the container. In the latter case the container would have
no sharding metadata, so if for some reason cleaving did not happen,
we'd end up with updates in an orphaned container.

Now, the sharder will repeatedly attempt to create shard containers
until successful, rather than leaving it to be a by-product of later
cleaving. That will at least reduce the variety of states we can
expect to find the shard containers in.

Change-Id: Ief5e28d95cdbc7afe5602ae83e3be4b0cc893faf