Override root object count with actual listing length when possible
When constructing a container listing from shards' listings, the
actual listing length may different from the root container's object
count if shard containers have not updated the root stats.

Before, the X-Container-Object-Count returned to the client was always
the value returned by the root container. With this patch that value
is overridden *when appropriate* by the actual listing length.  This
is only appropriate when the listing is unconstrained (no marker etc.)
and does not get capped by the listng length limit.  In practice this
will only apply to shrinking sharded containers, but also to test
conditions where this patch should make listing results more
predictable.

Change-Id: Ia91fbeefde37ea64bf50220636522f3dc64a6728