Make sharded shards report zero object stats
Before, when a container had been sharded, its object stats (as
included in the result of get_info()) were taken from the sum of its
shard ranges stats.

With this patchset this is only done if the container is the root
container.

Shards that have sharded have been shrunk, and their effective
object_count and bytes_used should be zero to avoid distorting the
sharded container account stats.

Change-Id: I766de4977e2b766d511545299777f06908a7c0c2