Make root sharding state survive container delete
If the root container is deleted then its sharding sysmeta is lost,
including the sharding epoch. But we only need the sysmeta sharding
epoch to signal a transition to sharding, and we shouldn't be
deleting a container that has just started sharding. It should
therefore be ok for a root container to be deleted, revived
and shard again.

Also fix some intermittent probe test failures due to non-idempotent
shrinking operations.

Change-Id: Id7a6380812928c1b7fa9f2fc3ef353dab405d3f6