Kill rsync coros when lockup detector tries to kill the process
Because the replicator in the master doesn't propergate the kill
signal to the subprocess in the coroutine. With the behavior, the lockup
detector causes a lot of rsync processes even it tries to reset the process.

This patch fixes the replicator kill rsync procs when the lockup detector
calls kill of eventlet threads.

Change-Id: Ifd14ce82de1f7ebb636d6131849e0fadb113a701