Track cleave progress per db file rather than per node
Before, the last range cleaved and cleave context was tracked per node
index. If the actual db being tracked did not make it to the same node
index during a rebalance then the tracking metadata would be incorrect
and could lead to rows not being cleaved.

With this change the cleaving progress is tracked against the retiring
db hash and epoch.

The cleaving progress metadata is expanded to include a flag to
indicate when cleaving of a db has completed. This means that if
cleaving is completed but the unlink of the retiring db fails when
moving to sharded state then the cleaving will not be repeated. Before,
the completion of cleaving was deduced from the upper bound of the
last cleaved range being '', but this is potentially confused with
deleted sysmeta values which also take value ''.

This change means that the separate sharding context that was used to check
that a db had not changed between sharding starting and completing is
no longer required.

Change-Id: Ibb7bbb7f40b33001e9a56a8bf7fc235ac05b13dd