Add assertions for cleaved broker own shard range state
A sharding container sets a shard range to ACTIVE state as soon as it
has been cleaved so that listings can be redirected to it.  The ACTIVE
state may then be replicated to other replicas of the sharding
container.

However, when cleaving and replicating a shard broker, the shard
broker's own shard range state should not be set to ACTIVE even if the
shard range being cleaved is ACTIVE. This is to avoid a sub-shard of a
shard container updating its state to ACTIVE in the root container
before the shard informs the root that *all* of its sub-shards have
become ACTIVE.

This patch adds assertions to verify this behaviour.

Change-Id: Ic00a0235728963ff64dc5e66c7cf6727f07b5513