Stop replicating object rows when container could be sharded
Add a new 'get_shard_ranges' RPC and have the container-replicator use
it to do a two-way sync of shard ranges on all successful syncs.
Expect (and ignore) 500s during upgrades.

Then, be willing to abort replication if we have shard ranges
locally. Instead, wait for the container-sharder to cleave rows to
their (hopefully final) destination, leaving just the sharded epoch DB.
If a handoff, this will be cleaned up on the next round of replication.

Change-Id: I289f558381d028b4d3129e4e51549d3f1a58dc2f