Make shard-info annotate lines with node index
Before, each line was annotated with the node id. Now each
line is annotated with the node index which aids identifying the
leader node.

Also add an entry for actual object count as well as apparent object
count. When a container is sharded the actual object count will be
misplaced objects whereas the apparent object count is the sum of
object counts reported from shard containers.

Change-Id: I8736590376a5f82348f8358448b724085820c761