Add EXPANDING state to ShardRange
Prematurely expanding an acceptor's namespace and bumping its
timestamp [1] was a terrible idea. The acceptor might send this
expanded namespace to other replicas of the root container before the
leader root has replicated the original acceptor namespace with the
same bumped timestamp. This results in the acceptor namespace being
expanded in the root before the donor has cleaved and before the donor
is deleted from the root!

The goal of expanding the acceptor's namespace was to prevent object
updates redirected from the donor during shrinking being considered as
misplaced.  This patch takes a different approach and introduces an
EXPANDING state. Now the acceptor's namespace is left unchanged until
shrinking completes, but its state is moved to EXPANDING. Then, when
shrinking completes, the donor forwards an updated acceptor shard
range to the acceptor which moves it back to ACTIVE state.

This also helps avoid a similar problem when shrinking the final shard
to the root: the root considers itself sharded while it has shard
ranges, so considers any object misplaced and was prone to moving
objects that the final shard range was in the process of cleaving to
the root. Now the root is also put into EXPANDING state when accepting
the final shrinking shard range.

Note: There is still further work needed to ensure that non-leader
roots learn that they are in EXPANDING state *before* any final shard
replica cleaves to them.

[1] Ie65cd4ce0abc98452163f6acdfad3cce5cdd216f

Change-Id: I43cb6285537a24163bfe1245fe15fe1f1393294b