Use state rather than subtlety to make shards of shards authoritative
In the root container it is important that sub-shards of a sharding
shard are not considered as being authoritative for the purposes of
listings before they have *all* been cleaved and the sharding shard is
simultaneously deleted.

However, it is useful for sub-shards to considered authoritative by a
sharding shard as soon as they are individually cleaved so that
listing requests to the shard can be redirected to them. Maintaining
this distinction between root and shard behaviour requires careful
control of how cleaved sub-shards are reported to the root.

Before, we controlled when sub-shards appeared ACTIVE to the root by
using a subtle distortion of the shard range state pushed to newly
cleaved sub-shards, and by not including other shard ranges in root
container updates while in sharding state.

Now, a newly cleaved shard of a shard is put in state CLEAVED until
all shards have been found, at which point all CLEAVED shards are
simultaneously moved to ACTIVE. This enables much simpler control over
what is considered as authoritative: CLEAVED shard ranges are not
authoritative in the root but CLEAVED shard ranges are considered
authoritative in shards.

In summary:

CLEAVED -> use when handling listings of shards but not roots;
   use for update redirection.

ACTIVE -> use when handling listings of shards and roots; use
  for update redirection.

Also, include other shard ranges when updating root...

Before, when a shard container made its periodic update
to the root, it only included its own shard range. Now it
also includes other shard ranges that it knows about, unless
it is shrinking.

This change ensures that when a donor shard has shrunk and
cleaved itself to an acceptor, the acceptor will update the
root with its own new expanded namespace (as before) and the
deleted donor shard range (which wasn't happening before).

Change-Id: I74df5085f999095a7d38444299abf12038e62d76