Cleanup swift/common/middleware/s3api/test dir
The scripts in the dir was used to setup functional tests
but they will be never used in swift upstream so let's cleanup
those scripts.

The configs for ceph-s3-test would be useful to make non-voting
gate in the swift upstream, this patch moves them into doc/s3api/conf
dir that is not included source code tree, anyway.

Change-Id: I608ed3a620478bc4ab2fdd8901ea987e356298c9