Require sufficient replication of each cleaved shard range
Before, when cleaving shard ranges, the sharder created local db's for
each shard range and spawned a thread to replicate them. There was no
check that the replication of each db was successful.

Now, each cleaved db is replicated in turn and cleaving does not
progress until each db has been sufficiently replicated. Replication
is considered sufficient if a quorum of replicas were sync'd. If the
sharding node happens to be one of the cleave db's primaries then that
counts as a sync'd replica.

For example, with a 3 replica container ring, replication is
considered sufficient if the cleaved db was sync'd with at least 2
other nodes, or if the cleaved db is on one of it's primaries and also
sync'd with one other node.

Change-Id: I1653ed69c51c9789e2ef19fae762a96a63139766